+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Teacher.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.TeacherDTO;
 import jakarta.persistence.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -21,29 +22,38 @@ public class Teacher {
     @Column(name = "teacher_surname")
     private String teacher_surname;
 
+    @Column(name = "teacher_DNI")
+    private String teacher_DNI;
+
     @OneToMany(mappedBy = "teacher")
     private List<Course> teacher_courses;
 
     public Teacher() {
 
     }
-    public Teacher(String teacher_name, String teacher_surname) {
+    public Teacher(String teacher_name, String teacher_surname, String teacher_DNI) {
         this.teacher_name = teacher_name;
         this.teacher_surname = teacher_surname;
+        this.teacher_DNI = teacher_DNI;
     }
 
-    public Teacher(String teacher_name, String teacher_surname, List<Course> courses) {
+    public Teacher(String teacher_name, String teacher_surname, String teacher_DNI, List<Course> courses) {
         this.teacher_name = teacher_name;
         this.teacher_surname = teacher_surname;
+        this.teacher_DNI = teacher_DNI;
         this.teacher_courses = courses;
     }
 
     public Teacher(TeacherDTO teacherDTO) {
         this.teacher_name = teacherDTO.getTeacherName();
         this.teacher_surname = teacherDTO.getTeacherSurname();
-        this.teacher_courses = teacherDTO.getCourses();
+        this.teacher_DNI = teacherDTO.getTeacherDNI();
+        //this.teacher_courses = teacherDTO.getCourses();
     }
 
+    public String getTeacherDNI() {
+        return teacher_DNI;
+    }
     public String getTeacherName() {
         return teacher_name;
     }
@@ -70,5 +80,12 @@ public class Teacher {
 
     public void addCourse(Course course) {
         teacher_courses.add(course);
+    }
+
+    public List<Course> getCourses() {
+        if(teacher_courses == null){
+            return new ArrayList<>();
+        }
+        return teacher_courses;
     }
 }