+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/TeacherService.java
@@ -35,4 +35,7 @@ public class TeacherService {
         return teacherRepository.findAll();
     }
 
+    public Teacher findTeacherByDNI(String teacherDni) {
+        return teacherRepository.findTeacherByDNI(teacherDni);
+    }
 }